GARGOYLE: FROTZ: Merge Err class into Processor
