+++ engines/gargoyle/frotz/processor.cpp
@@ -132,7 +132,7 @@ Opcode Processor::ext_opcodes[64] = {
 };
 
 Processor::Processor(OSystem *syst, const GargoyleGameDescription *gameDesc) : 
-		GlkInterface(syst, gameDesc), Errors(),
+		GlkInterface(syst, gameDesc),
 		_finished(0), _sp(nullptr), _fp(nullptr), _frameCount(0),
 		zargc(0), _decoded(nullptr), _encoded(nullptr), _resolution(0),
 		_randomInterval(0), _randomCtr(0), first_restart(true), script_valid(false),
@@ -180,6 +180,7 @@ Processor::Processor(OSystem *syst, const GargoyleGameDescription *gameDesc) :
 	Common::fill(&_stack[0], &_stack[STACK_SIZE], 0);
 	Common::fill(&zargs[0], &zargs[8], 0);
 	Common::fill(&_buffer[0], &_buffer[TEXT_BUFFER_SIZE], '\0');
+	Common::fill(&_errorCount[0], &_errorCount[ERR_NUM_ERRORS], 0);
 }
 
 void Processor::initialize() {
