+++ engines/gargoyle/frotz/processor_buffer.cpp
@@ -27,6 +27,42 @@
 namespace Gargoyle {
 namespace Frotz {
 
+const char *const Processor::ERR_MESSAGES[ERR_NUM_ERRORS] = {
+    "Text buffer overflow",
+    "Store out of dynamic memory",
+    "Division by zero",
+    "Illegal object",
+    "Illegal attribute",
+    "No such property",
+    "Stack overflow",
+    "Call to illegal address",
+    "Call to non-routine",
+    "Stack underflow",
+    "Illegal opcode",
+    "Bad stack frame",
+    "Jump to illegal address",
+    "Can't save while in interrupt",
+    "Nesting stream #3 too deep",
+    "Illegal window",
+    "Illegal window property",
+    "Print at illegal address",
+    "Illegal dictionary word length",
+    "@jin called with object 0",
+    "@get_child called with object 0",
+    "@get_parent called with object 0",
+    "@get_sibling called with object 0",
+    "@get_prop_addr called with object 0",
+    "@get_prop called with object 0",
+    "@put_prop called with object 0",
+    "@clear_attr called with object 0",
+    "@set_attr called with object 0",
+    "@test_attr called with object 0",
+    "@move_object called moving object 0",
+    "@move_object called moving into object 0",
+    "@remove_object called with object 0",
+    "@get_next_prop called with object 0"
+};
+
 void Processor::flush_buffer() {
 	/* Make sure we stop when flush_buffer is called from flush_buffer.
 	 * Note that this is difficult to avoid as we might print a newline
@@ -85,9 +121,71 @@ void Processor::print_char(zchar c) {
 	}
 }
 
+void Processor::print_string(const char *s) {
+	char c;
+
+	while ((c = *s++) != 0) {
+		if (c == '\n')
+			new_line();
+		else
+			print_char(c);
+	}
+}
+
+void Processor::print_long(uint value, int base) {
+	unsigned long i;
+	char c;
+
+	for (i = (base == 10 ? 1000000000 : 0x10000000); i != 0; i /= base) {
+		if (value >= i || i == 1) {
+			c = (value / i) % base;
+			print_char(c + (c <= 9 ? '0' : 'a' - 10));
+		}
+	}
+}
+
 void Processor::new_line()  {
 	flush_buffer();
 	stream_new_line();
+}
+
+void Processor::runtimeError(ErrorCode errNum) {
+	int wasfirst;
+
+	if (errNum <= 0 || errNum > ERR_NUM_ERRORS)
+		return;
+
+	if (_err_report_mode == ERR_REPORT_FATAL
+		|| (!_ignore_errors && errNum <= ERR_MAX_FATAL)) {
+		flush_buffer();
+		error(ERR_MESSAGES[errNum - 1]);
+		return;
+	}
+
+	wasfirst = (_errorCount[errNum - 1] == 0);
+	_errorCount[errNum - 1]++;
+
+	if ((_err_report_mode == ERR_REPORT_ALWAYS)
+		|| (_err_report_mode == ERR_REPORT_ONCE && wasfirst)) {
+		long pc;
+		GET_PC(pc);
+		print_string("Warning: ");
+		print_string(ERR_MESSAGES[errNum - 1]);
+		print_string(" (PC = ");
+		print_long(pc, 16);
+		print_char(')');
+
+		if (_err_report_mode == ERR_REPORT_ONCE) {
+			print_string(" (will ignore further occurrences)");
+		}
+		else {
+			print_string(" (occurence ");
+			print_long(_errorCount[errNum - 1], 10);
+			print_char(')');
+		}
+
+		new_line();
+	}
 }
 
 } // End of namespace Scott
